backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Caching==2.3.0
Flask-Cors==4.0.1
Flask-Mail==0.9.1
celery==5.4.0

redis==5.0.7
python-dotenv==1.0.1
fpdf2==2.7.9
